+++ integrations/jira/CHANGELOG.md
@@ -7,6 +7,54 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+## 0.9.0 (2026-09-07)
+
+
+### Features
+
+- Added actions support with a create_issue action for creating Jira issues
+
+
+## 0.8.5 (2026-09-07)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.51.1
+
+
+## 0.8.4 (2026-09-06)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.51.0
+
+
+## 0.8.3 (2026-09-06)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.50.11
+
+
+## 0.8.2 (2026-09-06)
+
+
+### Improvements
+
+- Use new core probe util to populate scopeless kinds
+
+
+## 0.8.1 (2026-09-06)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.50.10
+
+
 ## 0.8.0 (2026-09-03)